Restructure server package
