+++ .changeset/673706c0/changes.md
@@ -1,0 +1,1 @@
+- Restructure internal code
