+++ packages/server/index.js
@@ -1,1 +1,5 @@
-exports.WebServer = require('./WebServer');
+const WebServer = require('./lib/index');
+
+module.exports = {
+  WebServer,
+};
