+++ packages/server/lib/initConfig.js
